requirements.txt
markdown-it-py>=2.2.0
python-docx>=0.8.11
pytest>=7.3.1
black==23.0.0
pyinstaller==6.5.0  # 用于打包exe文件
requests>=2.28.2
html2docx>=1.6.0  # 用于HTML转换 